Add files via upload
This is mini project to manage a diagnostic center digitally. In this project there are 3 actors in the scene namely Doctor, Patient and Admin. Any patient can book an appointment for health checkup or any test online with the doctor of his choice. Doctor can delete or approve the appointment of any patient. Admin can see all the doctors and patients who have registered on the portal.